feat: 공모주 컨트롤러(IpoController)를 새로운 버전으로 교체
